requirements.txt
flask
pillow
python-telegram-bot
tweepy
redis
toml
webdavclient3